* Changed validations  for local file copying * Changed local_dir to support directories that don't exist yet * local_copy_dir now catches if an item with that same location exists * Added LocalDirectory and SFTPClient for typing purposes * made md5_hash public on all places for verification purposes #TODO Create list of destination generators

+++ mods/local_dir.py
@@ -14,8 +14,14 @@ class LocalDirectory:
         self.directory: Union[str, Path] = location
         if self.directory.is_dir():
             pass
+        elif self.directory.is_file():
+            print("Not a directory")
+            raise NotADirectoryError
+        elif not self.directory.exists():
+            pass
         else:
-            raise NotImplemented
+            print("Something is not right here, either make this non existent directory or an existing dir")
+            raise TypeError
         self.directory_items: List[Dict[str: str, str: LocalFile or LocalDirectory, str:str]] = list(
             {"name": item.name, "object": LocalFile(item) if item.is_file() else LocalDirectory(item),
              "type": "file" if item.is_file() else "directory"} for item in self.directory.rglob("*"))
@@ -26,7 +32,7 @@ class LocalDirectory:
     def __str__(self) -> str:
         return f"{str(self.directory)}"
 
-    def __md5_hash(self) -> str:
+    def md5_hash(self) -> str:
         """
         Get the md5 hash of your directory
         :return: str of your md5sum for the directory
@@ -39,7 +45,7 @@ class LocalDirectory:
         :param md5sum: str value of the md5 of another directory
         :return: True or False
         """
-        if self.__md5_hash() == md5sum:
+        if self.md5_hash() == md5sum:
             return True
         else:
             return False
@@ -56,18 +62,22 @@ class LocalDirectory:
         Copy the entire tree recursively to a local destination
         :return: Copied action
         """
-        try:
-            copytree(src=self.directory, dst=local_directory)
-            return True
-        except PermissionError as e:
-            print(f"You don't have permission to {self.__str__()}, please try with the right permissions")
-            return e
-        except FileNotFoundError as e:
-            print(f" {self.__str__()} file no longer exists or never did, please try again")
-            return e
-        except NotADirectoryError as e:
-            print(f" {self.__str__()} is a file and not a directory, please try again")
-            return e
+        if not local_directory.exists():
+            try:
+                copytree(src=self.directory, dst=local_directory)
+                return True
+            except PermissionError as e:
+                print(f"You don't have permission to {self.__str__()}, please try with the right permissions")
+                return e
+            except FileNotFoundError as e:
+                print(f" {self.__str__()} file no longer exists or never did, please try again")
+                return e
+            except NotADirectoryError as e:
+                print(f" {self.__str__()} is a file and not a directory, please try again")
+                return e
+        else:
+            print("Directory already exists, select a new location")
+            return IsADirectoryError
 
     def make_dir(self) -> True or False:
         """
@@ -81,15 +91,15 @@ class LocalDirectory:
             print(f"{self.__str__()} already exists, not creating directory")
             return False
 
-    def local_copy_file(self, file_name: str, destination: Union[str, Path]) -> True or Exception:
+    def local_copy_file(self, copy_file: LocalFile, destination: Union[str, Path]) -> True or Exception:
         """
-        Copy a file from this directory to anotherl ocation
-        :param file_name: Name of the file
+        Copy a file from this directory to another location
+        :param copy_file: Name of the file
         :param destination: Destination of the file
         :return: True or exception
         """
         for file in self.directory_items:
-            if file["name"] == file_name and file["type"] == "file":
+            if file["name"] == copy_file.file.name and file["object"].compare_md5(copy_file.md5_hash):
                 return file["object"].local_copy(destination)
             else:
                 print("This is not a file, it's a directory")
